Rewrite PluginParseURI & mbEncoderFind
PluginParseURI wasn't very human-readable, its for loop condition had
a value which changed at each iteration.

It is now rewritten to find where the pluginuid ends instead of
outputing the pluginuid and formatuid directly. This is easier to
understand and does not require to strdup the format URI.

There was a bug in mbEncoderFind which would compare a pluginuid (ie.
"king-pigeon") to a format URI (ie. "plugin://king-pigeon#smth"): with
correct values, this check cannot succeed. That means mbEncoderFind
always looped up to the last plugin in the registry and used this one.
This isn't fine when using more than one plugin.