Add global URI & context
An URI can be specified in the configuration at the same level as
`modbus` and `metadata`. If so, a global context is created from it at
binding startup and the connection is established. This global context
is used instead of a RTU-local one when an RTU configuration does not
specify an URI. This allows multiple serial RTUs to use the same
context (understand: the same serial link and connection).