Fix inconsistent answer after timeout
Previously, if a Modbus device answered after the timeout, the answer
would be stored in the read buffer (either TCP or serial) and would be
read at the next command sent to this device. That means the answer 1 -
was not up to date and 2 - potentially did not match the newer command.

The fix is to flush the read buffer matching the Modbus device before
sending a command. libmodbus' `modbus_flush` does that.

Redmine #6251