Replace polling frequency with polling period #6394
The new way to configure polling is with a period (which can be
specified in milliseconds, seconds and/or minutes). If a frequency is
specified (deprecated but still parsed for backwards compatibility),
it is converted to a period at parsing time. If both are specified,
crash.

Before that, the frequency was provided as an integer, which prevented
polling slower than once a second. Furthermore, afb_create_timer uses
a period in milliseconds, so a convertion occured.